LPS-86410 Resolve issue where deleting the Layout Prototype triggered another call to delete the prototype

+++ modules/apps/layout/layout-page-template-service/src/main/java/com/liferay/layout/page/template/service/impl/LayoutPageTemplateEntryLocalServiceImpl.java
@@ -205,8 +205,14 @@ public class LayoutPageTemplateEntryLocalServiceImpl
 		long layoutPrototypeId = layoutPageTemplateEntry.getLayoutPrototypeId();
 
 		if (layoutPrototypeId > 0) {
-			_layoutPrototypeLocalService.deleteLayoutPrototype(
-				layoutPrototypeId);
+			LayoutPrototype layoutPrototype =
+				_layoutPrototypeLocalService.fetchLayoutPrototype(
+					layoutPrototypeId);
+
+			if (layoutPrototype != null) {
+				_layoutPrototypeLocalService.deleteLayoutPrototype(
+					layoutPrototypeId);
+			}
 		}
 
 		// Resources
